i don't know why the default for this coulmn was an arry
