+++ src/email/mailer.ts
@@ -6,8 +6,6 @@ import path from "path";
 import { inspect } from "util";
 import logger from "../bootstrap/logger";
 import User from "../models/user";
-// import { config as dotenvConfig } from "dotenv";
-// dotenvConfig({path: path.resolve(__dirname, "../../.env")});
 
 export interface SendInBlueSendResponse {
     envelope: {
@@ -124,12 +122,3 @@ export const Emailer = {
 };
 
 Object.freeze(Emailer);
-
-async function test() {
-    const mailer = Emailer;
-    const user = new User({displayName: "Akosua", email: "tripleabatt@gmail.com"});
-    logger.info("BEFORE");
-    const res = await mailer.sendTestEmail(user);
-    logger.info(`RESULT: ${inspect(res)}`);
-}
-test();
