+++ tsconfig.json
@@ -64,5 +64,5 @@
     "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
     "emitDecoratorMetadata": true         /* Enables experimental support for emitting type metadata for decorators. */
   },
-  "exclude": ["./tests", "./dist", "./testSlackAnnouncer.ts"]
+  "exclude": ["./tests", "./dist", "./testSlackAnnouncer.ts", "./testMailer.ts"]
 }
